requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
#mysql-connector-python==8.0.34
mysql-connector-python==8.0.5
Werkzeug==3.0.4
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2